Add manual chart-releaser workflow
This change takes advantage of the "Helm Chart Releaser" in GitHub.
Information about this feature is located at:
https://github.com/marketplace/actions/helm-chart-releaser

Instead of being triggered on every push or on changes to the
chart itself, this action is configured to be invoked manually
as new releases are created.

As documented in the link above, for this feature to work, the
repository must have the `gh-pages` branch created and configured
in **Settings → Pages**:

https://github.com/marketplace/actions/helm-chart-releaser#pre-requisites

To create the required branch, consider using the following commands:

  git checkout --orphan gh-pages
  git rm -rf .
  git commit --allow-empty -m "Initialize gh-pages"
  git push origin gh-pages

Signed-off-by: Flavio Fernandes <ffernandes@nvidia.com>
(cherry picked from commit 6cb420dfd5089c463eb39d3d849f486b7404d534)